Move Cosmos extensions into a seperate file.
